refactor: move AppLayout
