+++ src/pages/teachers/key-stages/[keyStageSlug]/subjects/[subjectSlug]/programmes.tsx
@@ -3,7 +3,7 @@ import { GetStaticPathsResult, GetStaticProps, NextPage } from "next";
 
 import curriculumApi from "@/node-lib/curriculum-api";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
-import AppLayout from "@/components/AppLayout/AppLayout";
+import AppLayout from "@/components/SharedComponents/AppLayout";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import SubjectTierListing from "@/components/TeacherComponents/SubjectProgrammeListing";
 import {
